Extract panic hook from migration engine and port to others
