+++ introspection-engine/core/src/main.rs
@@ -11,6 +11,8 @@ use jsonrpc_core::*;
 use jsonrpc_stdio_server::ServerBuilder;
 
 fn main() {
+    user_facing_errors::set_panic_hook();
+
     let mut io_handler = IoHandler::new();
     io_handler.extend_with(RpcImpl::new().to_delegate());
