+++ libs/user-facing-errors/Cargo.toml
@@ -12,6 +12,7 @@ serde_json = "1.0.41"
 serde = { version = "1.0.102", features = ["derive"] }
 backtrace = "0.3.40"
 failure = "*"
+log = "0.4.8"
 
 [dependencies.quaint]
 git = "https://github.com/prisma/quaint.git"
