+++ libs/user-facing-errors/src/panic_hook.rs
@@ -1,0 +1,15 @@
+pub fn set_panic_hook() {
+    let original_hook = std::panic::take_hook();
+
+    std::panic::set_hook(Box::new(move |panic| {
+        let err = crate::UnknownError::new_in_panic_hook(&panic);
+
+        match serde_json::to_writer(std::io::stderr(), &err) {
+            Ok(_) => eprintln!(),
+            Err(err) => {
+                log::error!("Failed to write JSON error to stderr: {}", err);
+                original_hook(panic)
+            }
+        }
+    }));
+}
