+++ query-engine/prisma/src/main.rs
@@ -46,6 +46,8 @@ type AnyError = Box<dyn Error + Send + Sync + 'static>;
 
 #[tokio::main]
 async fn main() -> Result<(), AnyError> {
+    user_facing_errors::set_panic_hook();
+
     let matches = ClapApp::new("Prisma Query Engine")
         .version(env!("CARGO_PKG_VERSION"))
         .arg(
